backend/requirements.txt
Flask==2.2.5
Jinja2==3.1.2
gunicorn==20.1.0
line-bot-sdk==2.2.1
flask-login
flask-sqlalchemy==3.0.5
flask-cors
